Refactoring MegaDrone sample to use shared code
