+++ samples/shared/Oscillator.cpp
